+++ simple/src/main/java/com/ven/assists/simple/MainActivity.kt
@@ -35,7 +35,6 @@ import com.ven.assists.simple.overlays.OverlayStatusCard
 import com.ven.assists.simple.overlays.OverlayWeb
 import com.ven.assists.utils.ContactsUtil
 import com.ven.assists.utils.CoroutineWrapper
-import com.ven.assists.utils.NodeClassValue
 import com.ven.assists.window.AssistsWindowManager.overlayToast
 import kotlinx.coroutines.delay